Fix two more SDL2-era assumptions the port left behind
TTF_Init() was tested with "< 0". SDL2_ttf returned 0 on success and -1
on failure; SDL3_ttf returns a bool, so the test can never be true and a
failed init fell through to TTF_OpenFont instead of reporting itself.

The master-server query code stored tick counts in Uint32 while
SDL_GetTicks() returns Uint64. Truncating is self-consistent as long as
everything truncates, but "ping = SDL_GetTicks() - querystartticks"
mixes a full-width tick count with a 32-bit one, and yields a garbage
ping once the machine has been up for 49.7 days. Same defect class as
the NTimer widening earlier on this branch; widen these to match.

+++ src/NetPanzer/Views/MainMenu/Multi/MasterServer/ServerInfo.hpp
@@ -50,7 +50,9 @@ class ServerInfo {
   bool auth_on;
 
   network::Address ipaddress;
-  Uint32 querystartticks;
+  // See MSInfo::lastTicks -- must be as wide as SDL_GetTicks() returns, or
+  // the ping computed from it is nonsense after 49.7 days of uptime.
+  Uint64 querystartticks;
 
   int tryNum;
 };
